Waldtest {MCPAN}	R Documentation

Simultaneous Wald tests

Description

General function for adjusted p-values for an UIT in a one-way layout using multivariate normal distribution.

Usage


Waldtest(estp, varp, cmat, alternative = "greater")

Arguments

estp 	 numeric vector of point estimates of length I, with I = the number of samples

varp 	 numeric vector of variance estimates of length I, to be used for interval construction

cmat 	 Contrast matrix of dimension MxI, with M = the number of contrasts, I= the number of samples

alternative 	 character string

Details

Value

A list containing:

teststat 	 a numeric vector of teststatistics of length M

pval 	 a single numeric p-value, the p-value of the maximum test (minimum p-value)

p.val.adj 	 a vector of length M, the adjusted p-values of the single contrasts

alternative 	 a single character vector, as the input

Author(s)

Frank Schaarschmidt

See Also

For user level implementations see:

binomRDtest, poly3test

[Package MCPAN Index]