binomRDci {MCPAN}	R Documentation

Simultaneous confidence intervals for contrasts of independent binomial proportions (in a oneway layout)

Description

Simultaneous asymptotic CI for contrasts of binomial proportions, assuming that standard normal approximation holds. The contrasts can be interpreted as differences of (weighted averages) of proportions (risk ratios).

Usage



binomRDci(x,...)

## Default S3 method:
binomRDci(x, n, names=NULL,
 type="Dunnett", cmat=NULL, method="Wald",
 alternative="two.sided", conf.level=0.95, ...)

## S3 method for class 'formula':
binomRDci(formula, data,
 type="Dunnett", cmat=NULL, method="Wald",
 alternative="two.sided", conf.level=0.95,...)

## S3 method for class 'table':
binomRDci(x, type="Dunnett",
 cmat=NULL, method="Wald", alternative="two.sided",
 conf.level=0.95,...)

## S3 method for class 'matrix':
binomRDci(x, type="Dunnett",
 cmat=NULL, method="Wald", alternative="two.sided",
 conf.level=0.95,...)


Arguments

x 	 a numeric vector, giving the number of successes in I independent samples, or an object of class "table", representing the 2xk-table, or an object of class "matrix", representing the 2xk-table

n 	 a numeric vector, giving the number of trials (i.e. the sample size) in each of the I groups (only required if x is a numeric vector, ignored otherwise)

names 	 an optional character string, giving the names of the groups/ sample in x, n; if not specified the possible names of x are taken as group names (ignored if x is a table or matrix)

formula 	 a two-sided formula of the style 'response ~ treatment', where 'response' should be a categorical variable with two levels, while treatment should be a factor specifying the treatment levels

data 	 a data.frame, containing the variables specified in formula

type 	 a character string, giving the name of a contrast method, as defined in contrMat(multcomp); ignored if cmat is sepcified

cmat 	 a optional contrast matrix

method 	 a single character string, specifying the method for confidence interval construction; options are: "Wald", "ADD1", or "ADD2"

alternative 	 a single character string, one of "two.sided", "less", "greater"

conf.level 	 a single numeric value, simultaneous confidence level

... 	 arguments to be passed to binomest, currently only success labelling the event which should be considered as success

Details

See the examples for different usages.

Value

A object of class "binomRDci", a list containing:

conf.int 	 a matrix with 2 columns: lower and upper confidence bounds, and M rows

alternative 	 character string, as input

conf.level 	 single numeric value, as input

estimate 	 a matrix with 1 column: containing the estimates of the contrasts

x 	 the observed number of successes in the treatment groups

n 	 the number of trials in the treatment groups

p 	 the estimated proportions in the treatment groups

success 	 a character string labelling the event considered as success

names 	 the group names

method 	 a character string, specifying the method of interval construction

cmat 	 the contrast matrix used

Note

Note, that all implemented methods are approximate only. The coverage probability of the intervals might seriously deviate from the nominal level for small sample sizes and extreme success probabilities. See the simulation results in Sill (2007) for details.

Author(s)

Frank Schaarschmidt

References

Statistical procedures and characterization of coverage probabilities are described in: Sill, M. (2007): .... Master thesis, Institute of Biostatistics, Leibniz University Hannover.

Background references:

The ideas underlying the "ADD1" and "ADD2" adjustment are described in:

Agresti, A. and Caffo, B.(2000): Simple and effective confidence intervals for proportions and differences of proportions result from adding two successes and two failures. American Statistician 54, p. 280-288.

And have been generalized for a single contrast of I proportions in:

Price, R.M. and Bonett, D.G. (2004): An improved confidence interval for a linear function of binomial proportions. Computational Statistics and Data Analysis 45, 449-456.

See Also

summary.binomRDci, plot.sci

Examples



# In simple cases, counts of successes
# and number of trials can be just typed:

ntrials <- c(40,20,20,20)
xsuccesses <- c(1,2,2,4)
names(xsuccesses) <- LETTERS[1:4]
ex1D<-binomRDci(x=xsuccesses, n=ntrials, method="ADD1",
 type="Dunnett")
ex1D

ex1W<-binomRDci(x=xsuccesses, n=ntrials, method="ADD1",
 type="Williams", alternative="greater")
ex1W

# results can be plotted:
plot(ex1D, main="Comparisons to control group A")

# summary gives a more detailed print out:
summary(ex1W)

# if data are represented as dichotomous variable
# in a data.frame one can make use of table:

data(liarozole)

head(liarozole)

binomRDci(Improved ~ Treatment, data=liarozole, type="Tukey")
# here it might be important to define which level of the
# variable 'Improved' is to be considered as success
binomRDci(Improved ~ Treatment, data=liarozole, type="Tukey", success="y")

# If data are available as a named kx2-contigency table:

tab<-table(liarozole)
tab

binomRDci(tab, type="Tukey", success="y")


[Package MCPAN Index]