bronch {MCPAN}	R Documentation

Bronchial carcinoma data

Description

Data of a long term rodent carcinogenicity study of bronchial carcinoma.

Usage

data(bronch)

Format

A data frame with 200 observations on the following 3 variables.

group

a factor with levels 0 1 2 3, labelling the control and the three dose groups

Y

a logical vector, indicating whether a tumour was present at time of death (if TRUE), or not (if FALSE)

time

a numeric vector, the time of death, counted in days? from begin of the study

Details

Not yet checked for consistency with the source!

Source

...

References

...

Examples


data(bronch)
# raw tumour counts:

table(bronch[c("group","Y")])

# groupwise times of death:

boxplot(time ~ group, data=bronch, horizontal=TRUE)


[Package MCPAN Index]