censsample {MCPAN}	R Documentation

Random data for Poly-k

Description

Random numbers from two independent Weibull distributions for Mortality and tumour induction.

Usage


censsample(n, scale.m, shape.m, scale.t, shape.t = 3, tmax)

Arguments

n 	 a single numeric value, the number of individuals

scale.m 	 a single numeric value, scale parameter of the Weibull distribution for mortality

shape.m 	 a single numeric value, shape parameter of the Weibull distribution for mortality

scale.t 	 a single numeric value, scale parameter of the Weibull distribution for tumour induction

shape.t 	 a single numeric value, shape parameter of the Weibull distribution for tumour induction

tmax 	 a single numeric value, maximum time in the trial

Details

Value

A data.frame with columns

time 	 a numeric vector of length n, the time of death of an individual

status 	 a logical vector of length n, the tumour status at time of death (TRUE: tumour present, FALSE: no tumour present)

T.t 	 time of tumour induction (unobservable)

T.m 	 time of death

tmax 	 maximum time of death

Author(s)

Frank Schaarschmidt

References

Peddada (2005), Johnson, Kotz,...

[Package MCPAN Index]