corrMatgen {MCPAN}	R Documentation

A function to calculate the correlation matrix

Description

Correlation matrix for teststatistics and confidence intervals assuming multivariate standard normal distribution

Usage


corrMatgen(CM, varp)

Arguments

CM 	 a matrix of contrast coefficients, dimension MxI, where M=number of contrasts, and I=number of groups in a oneway layout

varp 	 a numeric vector of groupwise variance estimates (length = I)

Details

Value

A matrix of dimension MxM.

Author(s)

Frank Schaarschmidt

References

For correlation of contrasts of binomial proportion, see: Bretz F, Hothorn L.: Detecting dose-response using contrasts: asymptotic power and sample size determination for binomial data. Statistics in Medicine 2002; 21: 3325-3335.

[Package MCPAN Index]