plot.sci {MCPAN}	R Documentation

Plot simultaneous confidence intervals

Description

Creates a high level plot of confidence intervals from objects returned by functions binomRDci, binomORci, poly3ci. Plot function for objects of class "sci".

Usage


## S3 method for class 'sci':
plot(x, line = 0, linelty = 2, linelwd = 1, linecol = "black", CIvert = FALSE, CIlty = 1, CIlwd = 1, ...)
## S3 method for class 'binomRDci':
plot(x,...)
## S3 method for class 'poly3ci':
plot(x,...)
## S3 method for class 'binomORci':
plot(x, ...)

Arguments

x 	 an object of class "sci", "binomRDci", "poly3ci", as can be obtained by calling binomRDci,poly3ci

line 	 as single numeric value or vector, position of line(s) to be added orthogonal to the plotted CI

linelty 	 line type(s) for line as defined in par

linelwd 	 line width(s) for line as defined in par

linecol 	 colour(s) for line as defined in par

CIvert 	 logical value, specifying whether confidence intervals shall be represented by vertical (if TRUE) or horizontal (if FALSE) lines

CIlty 	 single value, line type for confidence intervals as defined in par

CIlwd 	 single value, line type for confidence intervals as defined in par

... 	 parameters to be passed to plot.default, as documented in plot.default, plot, or par

Details

Currently, the plot parameters ylim, xlim, las, axes, type, pch and the axis parameters labels, las are chosen internally by the function and can not be set by the user. Currently, combining several plots via par("mfrow") or layout in one device is not possible.

Value

A plot.

Author(s)

Frank Schaarschmidt

References

Examples



# An example with three dose groups
# compared to Placebo, binomial response:

resp<-c(2,8,6,13)
trials<-c(34,33,36,34)
names(resp)<-c("Placebo", "50", "75", "150")

ci<-binomRDci(x=resp, n=trials,
 type="Dunnett", method="ADD1")
plot(ci, line=c(0,0.1,0.2),linelty=c(1,2,3))


[Package MCPAN Index]