poly3ci {MCPAN}	R Documentation

Simultaneous confidence intervals for contrasts of poly-3-adjusted tumour rates

Description

Function to calculate simultaneous confidence intervals for several contrasts of poly-3-adjusted tumour rates in a oneway layout. Assuming a data situation as in Peddada(2005) or Bailer and Portier (1988). Simultaneous asymptotic CI for contrasts of tumour rates, assuming that standard normal approximation holds.

Usage


poly3ci(time, status, f, type = "Dunnett", cmat = NULL, method = "BP", alternative = "two.sided", conf.level = 0.95)

Arguments

time 	 a numeric vector of times of death of the individuals

status 	 a logical (or numeric, consisting of 0,1 only) vector giving the tumour status at time of death of each individual, where TRUE (1) = tumour present, FALSE (0) = no tumour present

f 	 a factor, giving the classification variable

type 	 a character string, giving the name of a contrast method, as defined in contrMat(multcomp)

cmat 	 a optional contrast matrix

method 	 a single charcter string, specifying the method for adjustment, with options: "BP" (Bailer Portier: assuming poly-3-adjusted rates are binomial variables), "BW" (Bieler, Williams: delta method as in Bieler and Williams (1993)) "ADD1" (as Bailer Portier, including an add1-adjustment on the raw tumour rates) "ADD2" (as Bailer Portier, including an add2-adjustment on the raw tumour rates following Agresti and Caffo (2000) for binomials)

alternative 	 a single character string

conf.level 	 a single numeric value, simultaneous confidence level

Details

Value

A object of class "poly3ci", a list containing:

conf.int 	 a matrix with 2 columns: lower and upper confidence bounds, and M rows

alternative 	 character string, as input

conf.level 	 single numeric value, as input

estimate 	 a numeric vector with the point estimates of the contrasts

time 	 as input

status 	 as input

f 	 as input

method 	 as input

cmat 	 as input, with colnames= factor levels of f

sample.est 	 a list containing sample estimates

Note

Please note that all methods here described are only approximative, and might violate the nominal level in certain situations. Please note further that appropriateness of the point estimates, and consequently of tests and confidence intervals is based on the assumptions in Bailer and Portier (1988), which might be a matter of controversies.

Author(s)

Frank Schaarschmidt

References

Assumption for poly-3-adjustment: Bailer, J.A. and Portier, C.J. (1988): Effects of treatment-induced mortality and tumor-induced mortality on tests for carcinogenicity in small samples. Biometrics 44, 417-431.

Peddada, S.D., Dinse, G.E., and Haseman, J.K. (2005): A survival-adjusted quantal response test for comparing tumor incidence rates. Applied Statistics 54, 51-61.

Bieler, G.S. and Williams, R.L. (1993): Ratio estimates, the Delta Method, and quantal response tests for increased carcinogenicity. Biometrics 49, 793-801.

Statistical procedures and characterization of coverage probabilities are described in: Sill, M. (2007): .... Master thesis, Institute of Biostatistics, Leibniz University Hannover.

Agresti, A. and Caffo, B.(2000): Simple and effective confidence intervals for proportions and differences of proportions result from adding two successes and two failures. American Statistician 54, p. 280-288.

See Also

Examples



data(methyl)
methylD<-poly3ci(time=methyl$death, status=methyl$tumour, f=methyl$group, type = "Dunnett", method = "ADD1" )
methylD
plot(methylD, main="Simultaneous CI for Poly-3-adjusted tumour rates")


[Package MCPAN Index]