poly3est {MCPAN}	R Documentation

Only for internal use.

Description

Poly-3- adjusted point and variance estimates for long term carcinogenicity data

Usage


poly3est(time, status, tmax, method = "BP", k=NULL)

Arguments

time 	 a numeric vector of times of death of the individuals

status 	 a logical (or numeric, consisting of 0,1 only) vector giving the tumour status at time of death of each individual, where TRUE (1) = tumour present, FALSE (0) = no tumour present

tmax 	 a single numeric vector, the final time of the trial

method 	 a single charcter string, specifying the method for adjustment, with options: "BP" (Bailer Portier: assuming poly-3-adjusted rates are binomial variables), "BW" (Bieler, Williams: delta method as in Bieler-Williams (1993)) "ADD1" (as Bailer Portier, including an add1-adjustment on the raw tumour rates) "ADD2" (as Bailer Portier, including an add2-adjustment on the raw tumour rates following Agresti Caffo (2000) for binomials)

k 	 a single numeric value, adjustment according to Bailer and Portier (1988), defaults to 3

Details

Only for internal use of poly3test and poly3ci.

Value

A list containing:

Y 	 number of tumours

n 	 number of individuals

estimate 	 poly-3-adjusted rates according to Bailer, Portier (1988)

weight 	 a vector of poly-3-adjusted weights, of length n

estp 	 poly-3-adjusted rate (according to method)

nadj 	 adjusted n (sum of weights)

varp 	 variance estimate (according to method)

varcor 	 variance estimate, if necessary corrected such that estimates of 0 can not occur

Note

Please note, that appropriateness of the point estimates seriously depends on whether the assumptions in Bailer and Portier are met or not.

Author(s)

Frank Schaarschmidt

References

Bailer, J.A. and Portier, C.J. (1988): Effects of treatment-induced mortality and tumor-induced mortality on tests for carcinogenicity in small samples. Biometrics 44, 417-431.

[Package MCPAN Index]