poly3estf {MCPAN}	R Documentation

Only for internal use.

Description

Poly-3- adjusted point and variance estimates for long term carcinogenicity data if data are given as a numeric time vector, a logical status vector and a factor containing a grouping variable

Usage


poly3estf(time, status, tmax=NULL, f, method = "BP", k=NULL)

Arguments

time 	 a numeric vector of times of death of the individuals

status 	 a logical (or numeric, consisting of 0,1 only) vector giving the tumour status at time of death of each individual, where TRUE (1) = tumour present, FALSE (0) = no tumour present

tmax 	 a single numeric value, the time of sacrifice in the trial, or the last last time of death, defaults to the maximal value observed in time

f 	 a factor of the same length as time status, giving the levels of a grouping variable in a one-way layout

method 	 a single charcter string, specifying the method for adjustment, with options: "BP" (Bailer Portier: assuming poly-3-adjusted rates are binomial variables), "BW" (Bieler, Williams: delta method as in Bieler-Williams (1993)) "ADD1" (as Bailer Portier, including an add1-adjustment on the raw tumour rates) "ADD2" (as Bailer Portier, including an add2-adjustment on the raw tumour rates following Agresti Caffo (2000) for binomials)

k 	 single numeric value, the adjustment parameter according to Bailer and Portier (1988), defaults to 3

Details

For internal use.

Value

A list containing:

Y 	 a numeric vector, groupwise number of tumours

n 	 a numeric vector, groupwise number of individuals

estimate 	 a numeric vector, groupwise poly-3-adjusted rates according to Bailer, Portier (1988)

weight 	 a numeric vector of poly-3-adjusted weights

estp 	 a numeric vector, groupwise poly-3-adjusted rate (according to method)

nadj 	 adjusted n (sum of weights)

varp 	 a numeric vector, groupwise variance estimate (according to method)

varcor 	 a numeric vector, groupwise variance estimate, if necessary corrected such that estimates of 0 can not occur

names 	 a character vector, the levels of the grouping variable f

k 	 a single numeric value, as input

Note

See poly3est

Author(s)

Frank Schaarschmidt

References

See poly3est

Examples



[Package MCPAN Index]