powermcp {MCPAN}	R Documentation

Approximative power calculation for multiple contrast tests

Description

Approximative power calculation for multiple contrast tests, based on normal approximation.

Usage


powermcp(TExpH1, alpha = 0.05, corrmatH1, alternative = "two.sided")

Arguments

TExpH1 	 numeric vector: the expectation of the test statistics under the alternative

alpha 	 a single numeric value: alpha-level, defaults to 0.05

corrmatH1 	 a numeric matrix, the correlation matrix of the teststatistics under the alternative, must have same dimensions as length of TExpH1

alternative 	 a single character string, specifying the alternative, one of "two.sided", "less", or "greater"

Details

Any-pair-power is calcualted, the power of a Union-Intersection-Test.

The probability, that any teststatistic exceeds the critical value is calculated from a central multivariate standard normal distribution. The appropriateness of the result strongly depends on the assumptions, that teststatistics are truely following a standard normal distributions, which might not be the case for small sample sizes.

Value

A single numeric value, the approximative power.

Author(s)

Frank Schaarschmidt

Examples



[Package MCPAN Index]