Waldci {MCPAN}	R Documentation

Simultaneous Wald confidence intervals

Description

General function for simultaneous CIs in a one-way layout using multivariate normal distribution.

Usage


Waldci(cmat, estp, varp, varcor, alternative = "two.sided", conf.level = 0.95)

Arguments

cmat 	 Contrast matrix of dimension MxI, with M = the number of contrasts, I= the number of samples

estp 	 numeric vector of point estimates of length I, with I = the number of samples

varp 	 numeric vector of variance estimates of length I, to be used for interval construction

varcor 	 numeric vector of variance estimates of length I

alternative 	 character string

conf.level 	 single numeric vector

Details

Mainly for internal use.

Value

A list containing:

conf.int 	 a matrix with 2 columns: lower and upper confidence bounds, and M rows

alternative 	 character string, as input

conf.level 	 single numeric value, as input

Author(s)

Frank Schaarschmidt

See Also

For user level implementations see: binomRDci, binomORci, poly3ci

[Package MCPAN Index]