poly3table {MCPAN}	R Documentation

Summarize long term carcinogenicity data

Description

Function to summarize data of long term carcinogenicity trials in a text format. Data are assumed to consist of (1) a dichotomous variable, defining whether the tumour of interest was present in an individual animal at time of death, and (2) a numeric variable containing the time of death of an individual animal, and (3) a grouping factor.

Usage


poly3table(time, status, f, tumour = NULL, symbol = "*")

Arguments

time 	 a numeric vector, containing the time of death of an individual

status 	 a logical (or dichotomous categorical) vector

f 	 a factor, specifying treatment groups

tumour 	 the value which status obtains if a tumour is present in an individual at time of death

symbol 	 symbol to indicate presence of tumour in the text representation

Value

A named list, containing a character string for each group

Author(s)

Frank Schaarschmidt

Examples


data(methyl)
methyl
poly3table(time=methyl$death, status=methyl$tumour,
 f=methyl$group, tumour = 1, symbol = "*")


[Package MCPAN Index]