scalechoice {MCPAN}	R Documentation

Only for internal use.

Description

How to define scale parameters of the Weibull distribution in the poly-k-setting

Usage


scalechoice(shape = 3, pm = 0.2, t = 1)

Arguments

shape 	 shape parameter (usually shape=3)

pm 	 a numeric value (0<p<1), probability that an individual dies before time t

t 	 a numeric value, time t

Value

a single numeric value, giving the scale parameter needed to fulfill pm(t) for the current definition of the Weibull disstribution.

Note

Scale parameter is defined differently than that in Peddada (2005)

Author(s)

Frank Schaarschmidt

References

See Also

pweibull

Examples



[Package MCPAN Index]