summary.binomORci {MCPAN}	R Documentation

Detailed print out for binomORci

Description

Produces a more detailed print out of objects of class "binomORci", including summary statistics, the used contrast matrix and the confidence intervals.

Usage


## S3 method for class 'binomORci':
summary(object, ...)

Arguments

object 	 an object of class "binomORci" as created by function binomORci

... 	 ...

Value

A print out.

Author(s)

Frank Schaarschmidt

Examples



x<-c(1,3,6,7,5)
n<-c(30,30,30,30,30)
names<-LETTERS[1:5]

ORD<-binomORci(x=x, n=n, names=names, type="Dunnett", alternative="greater")
summary(ORD)

ORW<-binomORci(x=x, n=n, names=names, type="Williams", alternative="greater")
summary(ORW)



[Package MCPAN Index]