summary.binomRDci {MCPAN}	R Documentation

Detailed print out for binomRDci

Description

Produces a more detailed print out of objects of class "binomRDci", including summary statistics, the used contrast matrix and the confidence intervals.

Usage


## S3 method for class 'binomRDci':
summary(object, ...)

Arguments

object 	 an object of class "binomRDci" as created by function binomRDci

... 	 further arguments to be passed to summary, currently only digits for rounding is supported

Value

A print out.

Author(s)

Frank Schaarschmidt

[Package MCPAN Index]