summary.binomRDtest {MCPAN}	R Documentation

Detailed print out for binomRDtest

Description

Produces a more detailed print out of objects of class "binomRDtest", including summary statistics, the used contrast matrix and the p-values.

Usage


## S3 method for class 'binomRDtest':
summary(object, ...)

Arguments

object 	 an object of class "binomRDtest" as created by function binomRDtest

... 	 further arguments to be passed to summary, currently only digits for rounding is supported

Value

A print out.

Author(s)

Frank Schaarschmidt

[Package MCPAN Index]